requirements.txt
numpy
scipy
matplotlib
scikit-learn
librosa
torchaudio
torch
umap-learn>=0.5.8
